pyproject.toml
[tool.poetry]
name = "syn-tool"
version = "0.1.0"
description = "CLI tool for SAP-Shopify synchronization"
authors = ["Syed Asim Ali <s.asim.ali.se@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.1.0"
rich = "^13.0.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"
loguru = "^0.7.0"
tenacity = "^8.2.0"

[tool.poetry.scripts]
syn = "syn_tool.cli:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
